+++ changes.md
@@ -3,6 +3,12 @@
 <details>
 <summary>Added Property(s)</summary>
 
+- added property `approvalRuleMode` to type `BusinessUnit`
+- added property `approvalRuleMode` to type `BusinessUnitDraft`
+- added property `approvalRuleMode` to type `Company`
+- added property `approvalRuleMode` to type `CompanyDraft`
+- added property `approvalRuleMode` to type `Division`
+- added property `approvalRuleMode` to type `DivisionDraft`
 - added property `attributedTo` to type `CreatedBy`
 - added property `attributedTo` to type `LastModifiedBy`
 - added property `metaTitle` to type `ProductTailoringCreatedMessage`
@@ -140,6 +146,8 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `BusinessUnitApprovalRuleMode`
+- added type `BusinessUnitChangeApprovalRuleModeAction`
 - added type `Attribution`
 - added type `AttributionSource`
 - added type `CustomerIndexingProgress`
@@ -148,10 +156,12 @@
 - added type `CustomerSearchIndexingStatusResponse`
 - added type `CustomerSearchRequest`
 - added type `CustomerSearchResult`
+- added type `BusinessUnitApprovalRuleModeChangedMessage`
 - added type `ProductTailoringImageAddedMessage`
 - added type `ProductTailoringImagesSetMessage`
 - added type `ProductVariantTailoringAddedMessage`
 - added type `ProductVariantTailoringRemovedMessage`
+- added type `BusinessUnitApprovalRuleModeChangedMessagePayload`
 - added type `ProductTailoringImageAddedMessagePayload`
 - added type `ProductTailoringImagesSetMessagePayload`
 - added type `ProductVariantTailoringAddedMessagePayload`